+++ src/main/java/datamodel/Repository.java
@@ -43,7 +43,7 @@ public class Repository implements Serializable {
 	/**
 	 * ID of the repository.
 	 */
-	private Integer id = null;
+	private Long id = null;
 
 	private RepositoryInternalMetrics repositoryInternalMetrics = new RepositoryInternalMetrics();
 	
@@ -81,7 +81,7 @@ public class Repository implements Serializable {
 	 * @param name Name of the repository
 	 * @param id  ID of the repository
 	 */
-	public Repository(String url, String name, Integer id) {
+	public Repository(String url, String name, Long id) {
 		setUrl(url);
 		setName(name);
 		setId(id);
@@ -133,7 +133,7 @@ public class Repository implements Serializable {
 	 * @author Miguel Ángel León Bardavío - mlb0029
 	 * @return the id
 	 */
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
@@ -143,7 +143,7 @@ public class Repository implements Serializable {
 	 * @author Miguel Ángel León Bardavío - mlb0029
 	 * @param id the id to set
 	 */
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
